requirements.txt
cffi
pyserial
Pillow
ruamel.yaml
wxPython
urwid
schedule
